backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.26.160
botocore==1.29.160
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==40.0.2
dj-database-url==2.0.0
Django==3.2.18
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.5.0
django-environ==0.10.0
django-ses==3.5.0
django-storages[boto3]==1.13.2
django-timezone-field==6.0
djangorestframework==3.12.2
drf-jwt==1.19.2
environ==1.0
factory-boy==3.3.0
Faker==19.1.0
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.1
multidict==6.0.4
openai==0.28.0
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
pyppeteer==1.0.2
pytest==7.4.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.29.0
s3transfer==0.6.1
sentry-sdk==1.27.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
whitenoise==6.4.0
yarl==1.9.2
zipp==3.17.0
